+++ src/main/java/ru/yandex/money/common/dbqueue/dao/PostgresQueueDao.java
@@ -10,6 +10,8 @@ import ru.yandex.money.common.dbqueue.settings.QueueLocation;
 
 import javax.annotation.Nonnull;
 import java.time.Duration;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
 import static java.util.Objects.requireNonNull;
@@ -22,10 +24,12 @@ import static java.util.Objects.requireNonNull;
  */
 public class PostgresQueueDao implements QueueDao {
 
+    private final Map<QueueLocation, String> enqueueSqlCache = new ConcurrentHashMap<>();
+    private final Map<QueueLocation, String> deleteSqlCache = new ConcurrentHashMap<>();
+    private final Map<QueueLocation, String> reenqueueSqlCache = new ConcurrentHashMap<>();
+
+    @Nonnull
     private final NamedParameterJdbcTemplate jdbcTemplate;
-    private final String enqueueSql;
-    private final String deleteTaskSql;
-    private final String reenqueueTaskSql;
     @Nonnull
     private final QueueTableSchema queueTableSchema;
 
@@ -37,34 +41,8 @@ public class PostgresQueueDao implements QueueDao {
      */
     public PostgresQueueDao(@Nonnull JdbcOperations jdbcTemplate,
                             @Nonnull QueueTableSchema queueTableSchema) {
-        requireNonNull(jdbcTemplate);
-
         this.queueTableSchema = requireNonNull(queueTableSchema);
-        this.jdbcTemplate = new NamedParameterJdbcTemplate(jdbcTemplate);
-        enqueueSql = "INSERT INTO %s(" +
-                queueTableSchema.getQueueNameField() + "," +
-                queueTableSchema.getPayloadField() + "," +
-                queueTableSchema.getNextProcessAtField() + "," +
-                queueTableSchema.getReenqueueAttemptField() + "," +
-                queueTableSchema.getTotalAttemptField() +
-                (queueTableSchema.getExtFields().isEmpty() ? "" :
-                        queueTableSchema.getExtFields().stream().collect(Collectors.joining(", ", ", ", ""))) +
-                ") VALUES " +
-                "(:queueName, :payload, now() + :executionDelay * INTERVAL '1 SECOND', 0, 0" +
-                (queueTableSchema.getExtFields().isEmpty() ? "" : queueTableSchema.getExtFields().stream()
-                        .map(field -> ":" + field).collect(Collectors.joining(", ", ", ", ""))) +
-                ") RETURNING id";
-
-        deleteTaskSql = "DELETE FROM %s WHERE " + queueTableSchema.getQueueNameField() +
-                " = :queueName AND id = :id";
-
-        reenqueueTaskSql = "UPDATE %s SET " + queueTableSchema.getNextProcessAtField() +
-                " = now() + :executionDelay * INTERVAL '1 SECOND', " +
-                queueTableSchema.getAttemptField() + " = 0, " +
-                queueTableSchema.getReenqueueAttemptField() +
-                " = " + queueTableSchema.getReenqueueAttemptField() + " + 1 " +
-                "WHERE id = :id AND " +
-                queueTableSchema.getQueueNameField() + " = :queueName";
+        this.jdbcTemplate = new NamedParameterJdbcTemplate(requireNonNull(jdbcTemplate));
     }
 
     @Override
@@ -81,7 +59,7 @@ public class PostgresQueueDao implements QueueDao {
         queueTableSchema.getExtFields().forEach(paramName -> params.addValue(paramName, null));
         enqueueParams.getExtData().forEach(params::addValue);
         return requireNonNull(jdbcTemplate.queryForObject(
-                String.format(enqueueSql, location.getTableName()), params, Long.class));
+                enqueueSqlCache.computeIfAbsent(location, this::createEnqueueSql), params, Long.class));
     }
 
 
@@ -89,7 +67,7 @@ public class PostgresQueueDao implements QueueDao {
     public boolean deleteTask(@Nonnull QueueLocation location, long taskId) {
         requireNonNull(location);
 
-        int updatedRows = jdbcTemplate.update(String.format(deleteTaskSql, location.getTableName()),
+        int updatedRows = jdbcTemplate.update(deleteSqlCache.computeIfAbsent(location, this::createDeleteSql),
                 new MapSqlParameterSource()
                         .addValue("id", taskId)
                         .addValue("queueName", location.getQueueId().asString()));
@@ -100,12 +78,45 @@ public class PostgresQueueDao implements QueueDao {
     public boolean reenqueue(@Nonnull QueueLocation location, long taskId, @Nonnull Duration executionDelay) {
         requireNonNull(location);
         requireNonNull(executionDelay);
-        int updatedRows = jdbcTemplate.update(String.format(reenqueueTaskSql, location.getTableName()),
+        int updatedRows = jdbcTemplate.update(reenqueueSqlCache.computeIfAbsent(location, this::createReenqueueSql),
                 new MapSqlParameterSource()
                         .addValue("id", taskId)
                         .addValue("queueName", location.getQueueId().asString())
                         .addValue("executionDelay", executionDelay.getSeconds()));
         return updatedRows != 0;
+    }
+
+    private String createEnqueueSql(@Nonnull QueueLocation location) {
+        return "INSERT INTO " + location.getTableName() + "(" +
+                (location.getIdSequence().map(ignored -> queueTableSchema.getIdField() + ",").orElse("")) +
+                queueTableSchema.getQueueNameField() + "," +
+                queueTableSchema.getPayloadField() + "," +
+                queueTableSchema.getNextProcessAtField() + "," +
+                queueTableSchema.getReenqueueAttemptField() + "," +
+                queueTableSchema.getTotalAttemptField() +
+                (queueTableSchema.getExtFields().isEmpty() ? "" :
+                        queueTableSchema.getExtFields().stream().collect(Collectors.joining(", ", ", ", ""))) +
+                ") VALUES " +
+                "(" + location.getIdSequence().map(seq -> "nextval('" + seq + "'), ").orElse("") +
+                ":queueName, :payload, now() + :executionDelay * INTERVAL '1 SECOND', 0, 0" +
+                (queueTableSchema.getExtFields().isEmpty() ? "" : queueTableSchema.getExtFields().stream()
+                        .map(field -> ":" + field).collect(Collectors.joining(", ", ", ", ""))) +
+                ") RETURNING " + queueTableSchema.getIdField();
+    }
+
+    private String createDeleteSql(@Nonnull QueueLocation location) {
+        return "DELETE FROM " + location.getTableName() + " WHERE " + queueTableSchema.getQueueNameField() +
+                " = :queueName AND " + queueTableSchema.getIdField() + " = :id";
+    }
+
+    private String createReenqueueSql(@Nonnull QueueLocation location) {
+        return "UPDATE " + location.getTableName() + " SET " + queueTableSchema.getNextProcessAtField() +
+                " = now() + :executionDelay * INTERVAL '1 SECOND', " +
+                queueTableSchema.getAttemptField() + " = 0, " +
+                queueTableSchema.getReenqueueAttemptField() +
+                " = " + queueTableSchema.getReenqueueAttemptField() + " + 1 " +
+                "WHERE " + queueTableSchema.getIdField() + " = :id AND " +
+                queueTableSchema.getQueueNameField() + " = :queueName";
     }
 
 }