+++ src/test/java/ru/yandex/money/common/dbqueue/internal/dao/CustomMssqlQueuePickTaskDaoTest.java
@@ -1,24 +1,32 @@
 package ru.yandex.money.common.dbqueue.internal.dao;
 
 import org.junit.BeforeClass;
-
-import ru.yandex.money.common.dbqueue.dao.BaseDaoTest;
-import ru.yandex.money.common.dbqueue.utils.QueueDatabaseInitializer;
+import ru.yandex.money.common.dbqueue.dao.MssqlQueueDao;
+import ru.yandex.money.common.dbqueue.internal.pick.MssqlQueuePickTaskDao;
+import ru.yandex.money.common.dbqueue.utils.MssqlDatabaseInitializer;
 
 /**
  * @author Oleg Kandaurov
  * @author Behrooz Shabani
  * @since 25.01.2020
  */
-public class CustomMssqlQueuePickTaskDaoTest extends MssqlQueuePickTaskDaoTest {
+public class CustomMssqlQueuePickTaskDaoTest extends QueuePickTaskDaoTest {
 
     @BeforeClass
     public static void beforeClass() {
-        QueueDatabaseInitializer.databaseType = QueueDatabaseInitializer.DatabaseType.MS;
-        BaseDaoTest.beforeClass();
+        MssqlDatabaseInitializer.initialize();
     }
 
     public CustomMssqlQueuePickTaskDaoTest() {
-        super(TableSchemaType.MS_CUSTOM);
+        super(new MssqlQueueDao(MssqlDatabaseInitializer.getJdbcTemplate(), MssqlDatabaseInitializer.CUSTOM_SCHEMA),
+                pickTaskSettings -> new MssqlQueuePickTaskDao(MssqlDatabaseInitializer.getJdbcTemplate(),
+                        MssqlDatabaseInitializer.CUSTOM_SCHEMA, pickTaskSettings),
+                MssqlDatabaseInitializer.CUSTOM_TABLE_NAME, MssqlDatabaseInitializer.CUSTOM_SCHEMA,
+                MssqlDatabaseInitializer.getJdbcTemplate(), MssqlDatabaseInitializer.getTransactionTemplate());
+    }
+
+    @Override
+    protected String currentTimeSql() {
+        return "SYSDATETIMEOFFSET()";
     }
 }