+++ src/test/java/ru/yandex/money/common/dbqueue/internal/dao/CustomPostgresQueuePickTaskDaoTest.java
@@ -1,23 +1,31 @@
 package ru.yandex.money.common.dbqueue.internal.dao;
 
 import org.junit.BeforeClass;
-
-import ru.yandex.money.common.dbqueue.dao.BaseDaoTest;
-import ru.yandex.money.common.dbqueue.utils.QueueDatabaseInitializer;
+import ru.yandex.money.common.dbqueue.dao.PostgresQueueDao;
+import ru.yandex.money.common.dbqueue.internal.pick.PostgresQueuePickTaskDao;
+import ru.yandex.money.common.dbqueue.utils.PostgresDatabaseInitializer;
 
 /**
  * @author Oleg Kandaurov
  * @since 12.10.2019
  */
-public class CustomPostgresQueuePickTaskDaoTest extends PostgresQueuePickTaskDaoTest {
+public class CustomPostgresQueuePickTaskDaoTest extends QueuePickTaskDaoTest {
 
     @BeforeClass
     public static void beforeClass() {
-        QueueDatabaseInitializer.databaseType = QueueDatabaseInitializer.DatabaseType.PG;
-        BaseDaoTest.beforeClass();
+        PostgresDatabaseInitializer.initialize();
     }
 
     public CustomPostgresQueuePickTaskDaoTest() {
-        super(TableSchemaType.PG_CUSTOM);
+        super(new PostgresQueueDao(PostgresDatabaseInitializer.getJdbcTemplate(), PostgresDatabaseInitializer.CUSTOM_SCHEMA),
+                pickTaskSettings -> new PostgresQueuePickTaskDao(PostgresDatabaseInitializer.getJdbcTemplate(),
+                        PostgresDatabaseInitializer.CUSTOM_SCHEMA, pickTaskSettings),
+                PostgresDatabaseInitializer.CUSTOM_TABLE_NAME, PostgresDatabaseInitializer.CUSTOM_SCHEMA,
+                PostgresDatabaseInitializer.getJdbcTemplate(), PostgresDatabaseInitializer.getTransactionTemplate());
+    }
+
+    @Override
+    protected String currentTimeSql() {
+        return "now()";
     }
 }